separate create class operation and validation
